Create 2 pass algorithm for macro issue:#918
This algorithm is used for identifying macro name and performing expansion.Pass 1 of the 2 pass algorithm is used to recognize macro definition and save it while Pass 2 of 2 pass algorithm is used to recognize macro call and perform macro expansion. In order to check the functioning of code i have written sample input in comment at the bottom of code before main().