Create N_queen.cpp
This is the N-Queen problem, which can also be solved by backtracking with bit-masking, this is an optimised approach than the normal backtracking approach.
In this case, we don’t need a function  for finding the correct placing of the queen in each case basically the safe place, instead we use bitsets which work in O(1) time. For rows and columns a simple array can be maintained for checking the available places while for the diagonals, the difference between the indices will be taken into consideration for checking the availability of a safe place.