Backtracking using Bitmasking
BacktrackingBacktracking is an algorithmic-technique for solving problems recursively by trying to build a solution incrementally, one piece at a time, removing those solutions that fail to satisfy the constraints of the problem at any point of time.  Here, we'll be taking the case of a backtracking based problem: N_queen problem. The N Queen is the problem of placing N chess queens on an N×N chessboard so that no two queens attack each other. We'll be implementing the solution using backtracking and bitmasking in various  programming languages.